Router installed, skeleton up
