+++ client/src/Main.css
@@ -1,0 +1,7 @@
+.Main {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 50px;
+}
