+++ client/src/components/Footer/Footer.css
@@ -4,6 +4,8 @@
   display: flex;
   justify-content: space-evenly;
   align-items: center;
+  position: absolute;
+  bottom: 0;
 }
 
 .to-top-btn{
@@ -15,3 +17,4 @@
   border-radius: 4px;
   border: 1px solid black;
 }
+
