+++ client/src/components/Team/Team.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Setting = props => {
+
+  return (
+  <div lassName="team">
+    <h1>Team Page</h1>
+  </div>
+  )
+}
+
+export default Setting;
